# internal/models/models.go
package models

import (
	"encoding/json"
	"time"
)

// Envelope is the standard top-level WebSocket frame
type Envelope struct {
	Type string          `json:"type"` // "chat", "server_ack", "delivery_ack", "read_ack", "sync_req", "sync_res", "presence_query", "presence_res", "error"
	Data json.RawMessage `json:"data"`
}

// InboundChatMessage is sent by the client when composing a message
type InboundChatMessage struct {
	ConversationID string `json:"conversation_id"`
	RecipientID    string `json:"recipient_id,omitempty"` // Optional for 1:1 direct chats
	ClientMsgID    string `json:"client_msg_id"`          // UUID generated by client for deduplication
	ContentType    string `json:"content_type,omitempty"` // Defaults to "text"
	Content        string `json:"content"`
}

// Message represents a fully sequenced, persisted chat message
type Message struct {
	ID             int64     `json:"id"`
	ConversationID string    `json:"conversation_id"`
	SeqID          int64     `json:"seq_id"`
	SenderID       string    `json:"sender_id"`
	ClientMsgID    string    `json:"client_msg_id"`
	ContentType    string    `json:"content_type"`
	Content        string    `json:"content"`
	CreatedAt      time.Time `json:"created_at"`
}

// ServerACK confirms the message was persisted, assigned a seq_id, and includes recipient reachability status
type ServerACK struct {
	ClientMsgID     string    `json:"client_msg_id"`
	MessageID       int64     `json:"message_id"`
	ConversationID  string    `json:"conversation_id"`
	SeqID           int64     `json:"seq_id"`
	Timestamp       time.Time `json:"timestamp"`
	RecipientStatus string    `json:"recipient_status"` // "online" | "offline"
	Delivered       bool      `json:"delivered"`        // true if sent to live recipient socket, false if queued offline
}

// DeliveryACK confirms recipient device received the message
type DeliveryACK struct {
	ConversationID string    `json:"conversation_id"`
	SeqID          int64     `json:"seq_id"`
	SenderID       string    `json:"sender_id"`
	RecipientID    string    `json:"recipient_id"`
	Timestamp      time.Time `json:"timestamp"`
}

// ReadACK confirms user opened the conversation and read up to seq_id (Watermark Cursor)
type ReadACK struct {
	ConversationID string    `json:"conversation_id"`
	SeqID          int64     `json:"seq_id"`
	SenderID       string    `json:"sender_id"`
	ReaderID       string    `json:"reader_id"`
	Timestamp      time.Time `json:"timestamp"`
}

// SyncRequest is sent by a client to catch up on offline/missed messages
type SyncRequest struct {
	ConversationID string `json:"conversation_id,omitempty"`
	RecipientID    string `json:"recipient_id,omitempty"`
	SinceSeqID     int64  `json:"since_seq_id"`
	Limit          int    `json:"limit,omitempty"`
}

// SyncResponse returns a paginated batch of missed messages
type SyncResponse struct {
	ConversationID string     `json:"conversation_id"`
	Messages       []*Message `json:"messages"`
	LatestSeqID    int64      `json:"latest_seq_id"`
	HasMore        bool       `json:"has_more"`
}

// PresenceQuery requests current online status of a user
type PresenceQuery struct {
	TargetUserID string `json:"target_user_id"`
}

// PresenceResponse returns current online/offline status
type PresenceResponse struct {
	UserID   string `json:"user_id"`
	Status   string `json:"status"` // "online" or "offline"
	LastSeen int64  `json:"last_seen,omitempty"`
}

// ErrorPayload sends descriptive error messages over WebSocket
type ErrorPayload struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// WSTicketResponse contains the short-lived single-use ticket for WebSocket authentication
type WSTicketResponse struct {
	Ticket    string `json:"ticket"`
	ExpiresIn int    `json:"expires_in"` // in seconds
}

// ConversationSummary represents a conversation preview in the conversation list API
type ConversationSummary struct {
	ID             string    `json:"id"`
	Type           string    `json:"type"`
	Title          string    `json:"title,omitempty"`
	RecipientID    string    `json:"recipient_id,omitempty"`
	LastMessage    *Message  `json:"last_message,omitempty"`
	UnreadCount    int64     `json:"unread_count"`
	LastSeqID      int64     `json:"last_seq_id"`
	LastReadSeqID  int64     `json:"last_read_seq_id"`
	UpdatedAt      time.Time `json:"updated_at"`
}

// CreateConversationRequest payload for creating a new conversation
type CreateConversationRequest struct {
	RecipientID string `json:"recipient_id"`
	Title       string `json:"title,omitempty"`
	Type        string `json:"type,omitempty"` // "direct" or "group"
}